Lesson 2.
Some differences are starting to show up, and I have a few questions.

I was not able to get the barycentric coordinates version of
draw_triangles working, which is kind of interesting. I can't pinpoint
what I was doing wrong that was resulting in every point within a bounding
box to contain a negative coordinate when performing the change of basis.

There's a few other questions scattered in comments. I really wish I had
someone to ask about this stuff. Hm.